chore: using client impl PIDM-545
