+++ src/main/java/it/gov/pagopa/authorizer/client/AuthCosmosClient.java
@@ -1,55 +1,16 @@
 package it.gov.pagopa.authorizer.client;
 
-import com.azure.cosmos.CosmosClient;
-import com.azure.cosmos.CosmosClientBuilder;
-import com.azure.cosmos.CosmosContainer;
-import com.azure.cosmos.CosmosDatabase;
-import com.azure.cosmos.models.CosmosQueryRequestOptions;
 import com.azure.cosmos.models.FeedResponse;
 import it.gov.pagopa.authorizer.entity.SubscriptionKeyDomain;
 
-public class AuthCosmosClient {
+public interface AuthCosmosClient {
 
-    private static AuthCosmosClient instance;
-
-    private final String databaseId = System.getenv("COSMOS_AUTH_DB_NAME");
-    private final String containerId = System.getenv("COSMOS_AUTH_CONTAINER_NAME");
-    private final int cosmosPageSize = Integer.parseInt(System.getenv().getOrDefault("COSMOS_PAGE_SIZE", "100"));
-
-    private final CosmosClient cosmosClient;
-
-    private AuthCosmosClient() {
-        String azureKey = System.getenv("COSMOS_AUTH_KEY");
-        String serviceEndpoint = System.getenv("COSMOS_AUTH_ENDPOINT");
-
-        this.cosmosClient = new CosmosClientBuilder()
-                .endpoint(serviceEndpoint)
-                .key(azureKey)
-                .buildClient();
-    }
-
-    public AuthCosmosClient(CosmosClient cosmosClient) {
-        this.cosmosClient = cosmosClient;
-    }
-
-    public static AuthCosmosClient getInstance() {
-        if (instance == null) {
-            instance = new AuthCosmosClient();
-        }
-        return instance;
-    }
-
-    public Iterable<FeedResponse<SubscriptionKeyDomain>> getSubkeyDomainPage(String domain, String continuationToken) {
-        CosmosDatabase cosmosDatabase = this.cosmosClient.getDatabase(databaseId);
-        CosmosContainer cosmosContainer = cosmosDatabase.getContainer(containerId);
-
-        //Build query
-        String query = String.format("SELECT * FROM SubscriptionKeyDomain s WHERE s.domain = '%s'",
-                domain);
-
-        //Query the container
-        return cosmosContainer
-                .queryItems(query, new CosmosQueryRequestOptions(), SubscriptionKeyDomain.class)
-                .iterableByPage(continuationToken, cosmosPageSize);
-    }
+    /**
+     * Sends a request to refresh the authorization configuration using the given parameters.
+     *
+     * @param domain the domain for which the configuration is to be refreshed.
+     * @param continuationToken continuation token for the pagination implementation.
+     * @return the iterable containing the selected pages by the query.
+     */
+    Iterable<FeedResponse<SubscriptionKeyDomain>> getSubkeyDomainPage(String domain, String continuationToken);
 }
